+++ cr_payment_eupago/doc/changelog.rst
@@ -1,8 +1,10 @@
-18.0.0.0(Date: 28th July, 2026)
+19.0.0.0(Date: 24th July, 2026)
 -------------------------------
-- Migrated Module
 - Supports Multibanco (ATM Reference) payments directly in Odoo.
 - Enables MB WAY transactions for instant mobile payments.
 - Supports 3D Secure Credit Card payments for enhanced security.
 - Automated synchronization of payment statuses with Odoo.
 - Added API based refund for Credit Card and MB WAY payments.
+- Added dynamic payment fees calculation (Fixed/Percentage) for Domestic and International transactions.
+- Enabled fees configuration for all payment methods of euPago
+- Added Fees badge in payment form to the checkout page.